% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql_session.R
\name{cql_connect}
\alias{cql_connect}
\title{Cassandra Session}
\usage{
cql_connect(sc, cluster, keyspace)
}
\arguments{
\item{sc}{Spark context used for the scala context}

\item{cluster}{The target cluster to which to connect}

\item{keyspace}{The keyspace to which to connect}
}
\description{
Return a connection session
}
